Update `Issues` GraphQL query and related codes
- Added new fields (specified in #169) to the GraphQL query.
- Restored previously commented-out project-related fields.
- Explicitly mapped GraphQL custom scalar `URI` to Rust.
- Updated `GitHubIssue` struct to align with the `Issue` struct, and
  adjusted related functions accordingly.
- Made alias for enums related with `GitHubIssue` struct.
- Changed the numeric type in `parse_key`'s return value from `i64` to
  `i32` to align with the GraphQL API.

Close: #169

+++ src/database.rs
@@ -151,7 +151,7 @@ impl<T: TryFromKeyValue> DoubleEndedIterator for Iter<T> {
     }
 }
 
-pub(crate) fn parse_key(key: &[u8]) -> Result<(String, String, i64)> {
+pub(crate) fn parse_key(key: &[u8]) -> Result<(String, String, i32)> {
     let re = Regex::new(r"(?P<owner>\S+)/(?P<name>\S+)#(?P<number>[0-9]+)").expect("valid regex");
     if let Some(caps) = re.captures(
         String::from_utf8(key.to_vec())
@@ -172,7 +172,7 @@ pub(crate) fn parse_key(key: &[u8]) -> Result<(String, String, i64)> {
             .name("number")
             .ok_or_else(|| anyhow!("invalid key"))?
             .as_str()
-            .parse::<i64>()
+            .parse::<i32>()
             .context("invalid key")?;
         Ok((owner, name, number))
     } else {
